Adjust SetsAndElements integration test
Added new case attempting to create a new SetElement pointing to a
node used in an already added SetElement. API_EEXIST expected

Adjusted expected error codes for INVALID_HANDLE / UNDEF used in bulk
addition and removal

Add support for extended error codes in `aepb` and `aerb` SetElement

Received API_EEXIST error may now be received in an extended format which
requires adjustment

+++ src/commands.cpp
@@ -9359,6 +9359,37 @@ bool CommandSE::procerrorcode(const Result& r, Error& e) const
     return false;
 }
 
+bool CommandSE::procExtendedError(int64_t& errCode, handle& eid) const
+{
+    bool isErr = false;
+    for (;;)
+    {
+        switch (client->json.getnameid())
+        {
+        case MAKENAMEID3('e', 'r', 'r'):
+        {
+            isErr = true;
+            errCode = client->json.getint();
+            break;
+        }
+
+        case MAKENAMEID3('e', 'i', 'd'):
+        {
+            eid = client->json.gethandle(MegaClient::SETELEMENTHANDLE);
+            break;
+        }
+
+        default:
+            if (!client->json.storeobject()) return false;
+            break;
+
+        case EOO:
+            return isErr;
+        }
+    }
+    return isErr;
+}
+
 CommandPutSet::CommandPutSet(MegaClient* cl, Set&& s, unique_ptr<string> encrAttrs, string&& encrKey,
                              std::function<void(Error, const Set*)> completion)
     : mSet(new Set(move(s))), mCompletion(completion)
@@ -9568,6 +9599,7 @@ bool CommandPutSetElements::procresult(Result r)
     bool allOk = true;
     vector<const SetElement*> addedEls;
     vector<int64_t> errs(mElements->size(), API_OK);
+    handle errEid = UNDEF;
     for (size_t elCount = 0u; elCount < mElements->size(); ++elCount)
     {
         if (client->json.isnumeric())
@@ -9577,21 +9609,31 @@ bool CommandPutSetElements::procresult(Result r)
         }
         else if (client->json.enterobject())
         {
-            handle elementId = 0;
-            m_time_t ts = 0;
-            int64_t order = 0;
-            if (!procjsonobject(elementId, ts, nullptr, nullptr, nullptr, &order))
+            const auto posAux = client->json.pos;
+            if (procExtendedError(errs[elCount], errEid))
             {
-                LOG_err << "Sets: failed to parse Element object in `aepb` response";
-                allOk = false;
-                break;
+                if (errEid == UNDEF) LOG_warn << "Sets: Extended error missing Element id";
+                errEid = UNDEF;
             }
+            else
+            {
+                client->json.pos = posAux;
+                handle elementId = 0;
+                m_time_t ts = 0;
+                int64_t order = 0;
+                if (!procjsonobject(elementId, ts, nullptr, nullptr, nullptr, &order))
+                {
+                    LOG_err << "Sets: failed to parse Element object in `aepb` response";
+                    allOk = false;
+                    break;
+                }
 
-            SetElement& el = mElements->at(elCount);
-            el.setId(elementId);
-            el.setTs(ts);
-            el.setOrder(order);
-            addedEls.push_back(client->addOrUpdateSetElement(move(el)));
+                SetElement& el = mElements->at(elCount);
+                el.setId(elementId);
+                el.setTs(ts);
+                el.setOrder(order);
+                addedEls.push_back(client->addOrUpdateSetElement(move(el)));
+            }
 
             if (!client->json.leaveobject())
             {
@@ -9705,6 +9747,7 @@ CommandRemoveSetElements::CommandRemoveSetElements(MegaClient* cl, handle sid, v
 
 bool CommandRemoveSetElements::procresult(Result r)
 {
+    bool allOk = true;
     Error e = API_OK;
     if (procerrorcode(r, e))
     {
@@ -9727,7 +9770,33 @@ bool CommandRemoveSetElements::procresult(Result r)
     vector<int64_t> errs(mElemIds.size());
     for (size_t elCount = 0u; elCount < mElemIds.size(); ++elCount)
     {
-        errs[elCount] = client->json.getint();
+        if (client->json.isnumeric())
+        {
+            errs[elCount] = client->json.getint();
+        }
+        else if (client->json.enterobject())
+        {
+            handle errEid = UNDEF;
+            if (procExtendedError(errs[elCount], errEid))
+            {
+                if (errEid == UNDEF) LOG_warn << "Sets: Extended error missing Element id in `aerb`";
+            }
+            else allOk = false;
+
+            if (!client->json.leaveobject())
+            {
+                LOG_err << "Sets: failed to parse Element object in `aerb` response";
+                allOk = false;
+            }
+        }
+        else
+        {
+            LOG_err << "Sets: failed to parse Element removal response in `aerb` command response";
+            allOk = false;
+        }
+
+        if (!allOk) break;
+
         if (errs[elCount] == API_OK && !client->deleteSetElement(mSetId, mElemIds[elCount]))
         {
             LOG_err << "Sets: Failed to remove Element in `aerb` command response";
@@ -9740,7 +9809,7 @@ bool CommandRemoveSetElements::procresult(Result r)
         mCompletion(e, &errs);
     }
 
-    return true;
+    return allOk;
 }
 
 CommandRemoveSetElement::CommandRemoveSetElement(MegaClient* cl, handle sid, handle eid, std::function<void(Error)> completion)
